# build/swagger.yaml
openapi: 3.0.0
info:
    title: ec2
    contact: {}
    version: '1.0'
servers:
    - url: https://newsapp-cf64.onrender.com/api
      variables: {}
paths:
...
    /auth/token_login:
        post:
            tags:
                - auth
            summary: token_login
            operationId: token_login
            parameters: []
            responses:
                '200':
                    description: ''
                    headers: {}
            deprecated: false
            security:
                - bearer: []
    /auth/refresh_tokens:
        post:
            tags:
                - auth
            summary: refresh_tokens
            operationId: refresh_tokens
            parameters: []
            responses:
                '200':
                    description: ''
                    headers: {}
            deprecated: false
            security:
                - bearer: []
...
components:
...
    securitySchemes:
        bearer:
            type: http
            scheme: bearer
security: []
...